fix(ChartComponent): renamed component
