+++ src/app/components/chart-component/chart.component.scss
